+++ pdata/pmetric/pmetricotlp/metrics.go
@@ -65,6 +65,7 @@ func (mr Response) UnmarshalJSON(data []byte) error {
 }
 
 // Request represents the request for gRPC/HTTP client/server.
+// It's a wrapper for pmetric.Metrics data.
 type Request struct {
 	orig *otlpcollectormetrics.ExportMetricsServiceRequest
 }
@@ -72,6 +73,13 @@ type Request struct {
 // NewRequest returns an empty Request.
 func NewRequest() Request {
 	return Request{orig: &otlpcollectormetrics.ExportMetricsServiceRequest{}}
+}
+
+// NewRequestFromMetrics returns a Request from pmetric.Metrics.
+// Because Request is a wrapper for pmetric.Metrics,
+// any changes to the provided Metrics struct will be reflected in the Request and vise versa.
+func NewRequestFromMetrics(m pmetric.Metrics) Request {
+	return Request{orig: internal.MetricsToOtlp(m)}
 }
 
 // MarshalProto marshals Request into proto bytes.
@@ -102,6 +110,7 @@ func (mr Request) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// Deprecated: [v0.50.0] Use NewRequestFromMetrics instead.
 func (mr Request) SetMetrics(ld pmetric.Metrics) {
 	*mr.orig = *internal.MetricsToOtlp(ld)
 }